Controllo token e task per titolo

+++ factories/TaskFactory.php
@@ -38,12 +38,37 @@ abstract class TaskFactory
         return $response;
     }
 
-    public static function getListTaskList($listId)
+    public static function getListByTitle($title) {
+        $client = GoogleHelper::getClient();
+        $service = new \Google_Service_Tasks($client);
+        $results = $service->tasklists->listTasklists(['maxResults'=> 0 ]);
+        foreach ($results->getItems() as $tasklist) {
+            if(trim($tasklist->getTitle()) == $title) {
+                return new Task([
+                    'id' => $tasklist->getId(),
+                    'etag' => $tasklist->getEtag(),
+                    'title' => $tasklist->getTitle()
+                ]);
+            }
+        }
+    }
+
+    public static function getListTaskList($listId, $request = [])
     {
+        if(!$listId && isset($request['title'])) {
+            $list = TaskFactory::getListByTitle($request['title']);
+            if(!$list) {
+                $response['status_code_header'] = 'HTTP/1.1 409';
+                $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+                return $response;
+            }
+            $response['status_code_header'] = 'HTTP/1.1 200 OK';
+            $response['body'] = json_encode($list);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasklists->get($listId);
-        $response['status_code_header'] = 'HTTP/1.1 200 OK';
         $response['body'] = json_encode(new Task([
             'id' => $results->getId(),
             'etag' => $results->getEtag(),
@@ -70,8 +95,19 @@ abstract class TaskFactory
 
     public static function updateTaskList($listId, $data = [])
     {
+        if(!$listId && !isset($data['id']) && isset($data['listTitle'])) {
+            $list = TaskFactory::getListByTitle($data['listTitle']);
+            if($list) $data['id'] = $list->id;
+        }
+        $listId = $data['id'];
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
+        $listId = $data['id'];
         $postData = new \Google_Service_Tasks_TaskList($data);
         $results = $service->tasklists->update($listId, $postData);
 
@@ -84,8 +120,17 @@ abstract class TaskFactory
         return $response;
     }
 
-    public static function deleteTaskList($listId)
+    public static function deleteTaskList($listId, $request = [])
     {
+        if(!$listId && isset($request['title'])) {
+            $list = TaskFactory::getListByTitle($request['title']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasklists->delete($listId);
@@ -95,17 +140,28 @@ abstract class TaskFactory
     }
 
 
-    public static function getTaskLists($listId)
+    public static function getTaskLists($listId, $request = [])
     {
+        if(!$listId && isset($request['listTitle'])) {
+            $list = TaskFactory::getListByTitle($request['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasks->listTasks($listId);
+        
         $tasksArray = array();
         foreach ($results->getItems() as $task) {
             $tasksArray[] = new Task([
                 'id' => $task->id,
                 'title' => $task->title,
                 'etag' => $task->etag,
+                'due' => $task->due,
                 'completed' => $task->completed,
                 'deleted' => $task->deleted,
                 'hidden' => $task->hidden,
@@ -120,17 +176,54 @@ abstract class TaskFactory
         return $response;
     }
 
-    public static function getTask($listId, $taskId)
+    public static function getTaskByTitle($listId, $title) {
+        $client = GoogleHelper::getClient();
+        $service = new \Google_Service_Tasks($client);
+        $results = $service->tasks->listTasks($listId);
+        foreach ($results->getItems() as $task) {
+            if(trim($task->getTitle()) == $title) {
+                return new Task([
+                    'id' => $task->id,
+                    'title' => $task->title,
+                    'etag' => $task->etag,
+                    'due' => $task->due,
+                    'completed' => $task->completed,
+                    'deleted' => $task->deleted,
+                    'hidden' => $task->hidden,
+                    'notes' => $task->notes,
+                    'status' => $task->status,
+                    'updated' => $task->updated,
+                ]);
+            }
+        }
+    }
+
+    public static function getTask($listId, $taskId, $request = [])
     {
+        $response['status_code_header'] = 'HTTP/1.1 200 OK';
+        if(!$listId && isset($request['listTitle'])) {
+            $list = TaskFactory::getListByTitle($request['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
+        if((!$taskId || strlen($taskId) < 3) && isset($request['title'])) {
+            $task = TaskFactory::getTaskByTitle($listId, $request['title']);
+            $response['body'] = json_encode($task);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasks->get($listId, $taskId);
 
-        $response['status_code_header'] = 'HTTP/1.1 200 OK';
         $response['body'] = json_encode(new Task([
             'id' => $results->id,
             'title' => $results->title,
             'etag' => $results->etag,
+            'due' => $task->due,
             'completed' => $results->completed,
             'deleted' => $results->deleted,
             'hidden' => $results->hidden,
@@ -143,6 +236,15 @@ abstract class TaskFactory
 
     public static function createTask($listId, $data = [])
     {
+        if(!$listId && isset($data['listTitle'])) {
+            $list = TaskFactory::getListByTitle($data['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $postData = new \Google_Service_Tasks_Task($data);
@@ -153,6 +255,7 @@ abstract class TaskFactory
             'id' => $results->id,
             'title' => $results->title,
             'etag' => $results->etag,
+            'due' => $results->due,
             'completed' => $results->completed,
             'deleted' => $results->deleted,
             'hidden' => $results->hidden,
@@ -165,6 +268,25 @@ abstract class TaskFactory
 
     public static function updateTask($listId, $taskId, $data)
     {
+        if(!$listId && isset($data['listTitle'])) {
+            $list = TaskFactory::getListByTitle($data['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
+        if(!$taskId && isset($data['taskTitle']) && !isset($data['id'])) {
+            $task = TaskFactory::getTaskByTitle($listId, $data['taskTitle']);
+            if($task) $data['id'] = $task->id;
+        }
+        $taskId = $data['id'];
+        if (!$taskId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Task non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $postData = new \Google_Service_Tasks_Task($data);
@@ -175,6 +297,7 @@ abstract class TaskFactory
             'id' => $results->id,
             'title' => $results->title,
             'etag' => $results->etag,
+            'due' => $task->due,
             'completed' => $results->completed,
             'deleted' => $results->deleted,
             'hidden' => $results->hidden,
@@ -185,23 +308,50 @@ abstract class TaskFactory
         return $response;
     }
 
-    public static function deleteTask($listId, $taskId)
+    public static function deleteTask($listId, $taskId, $request = [])
     {
+        if(!$listId && isset($request['listTitle'])) {
+            $list = TaskFactory::getListByTitle($request['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
+        if((!$taskId || strlen($taskId) < 5) && isset($request['title'])) {
+            $task = TaskFactory::getTaskByTitle($listId, $request['title']);
+            if($task) $taskId = $task->id;
+        }
+        if (!$taskId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Task non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasks->delete($listId, $taskId);
         $response['status_code_header'] = 'HTTP/1.1 200 OK';
-        $response['body'] = json_encode($tasksArray);
+        $response['body'] = json_encode($results);
         return $response;
     }
 
-    public static function clearListTasks($listId)
+    public static function clearListTasks($listId, $request)
     {
+        if(!$listId && isset($request['listTitle'])) {
+            $list = TaskFactory::getListByTitle($request['listTitle']);
+            if($list) $listId = $list->id;
+        }
+        if (!$listId) {
+            $response['status_code_header'] = 'HTTP/1.1 409';
+            $response['body'] = json_encode(["result" => false, "message" => "Lista non trovata"]);
+            return $response;
+        }
         $client = GoogleHelper::getClient();
         $service = new \Google_Service_Tasks($client);
         $results = $service->tasks->clear($listId);
         $response['status_code_header'] = 'HTTP/1.1 200 OK';
-        $response['body'] = json_encode($tasksArray);
+        $response['body'] = json_encode($results);
         return $response;
     }
 }
